chore(mcp): extract v1/v2 protocol handlers into separate files
Extract protocol-specific logic from cdpRelay into pluggable
ExtensionProtocolHandler implementations (cdpRelayV1, cdpRelayV2).
Rename env variable to PLAYWRIGHT_EXTENSION_PROTOCOL.
